requirements.txt
Flask
flask-cors
beautifulsoup4
google-cloud-translate
gunicorn